Move version_retention_period to backup_sample
